+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - added `symcat` in macros to concatenate a symbol and a number/string/symbol to create a new one
 - added `argcount` in macros to count (at compile time) the number of arguments of a function
 - fixed a bug where `(bloc)` and `(print bloc)`, given a `!{bloc value}` macro, didn't give the same result (one was applied, the other was partial)
+- internal hidden keyword `Pop` to use in the stack trashing
 
 ### Changed
 - updating doxyfile and some docstrings